[TEDIX] registration form
1) move School and City models to the tedix_ro app
2) add new form fields for student and teacher registration
3) show/hide student's form fields depending on Role

- https://youtrack.raccoongang.com/issue/T-17
- https://youtrack.raccoongang.com/issue/T-15
- https://youtrack.raccoongang.com/issue/T-18
