+++ common/djangoapps/tedix_ro/forms.py
@@ -2,10 +2,11 @@ from django import forms
 from django.contrib.auth.models import User
 from django.forms import ModelForm
 
-from tedix_ro.models import StudentParent
-
+from tedix_ro.models import StudentParent, City, School
 
 CLASSROOM_CHOICES = (
+    ('', ''),
+
     ('7A', '7A'),
     ('7B', '7B'),
     ('7C', '7C'),
@@ -17,19 +18,27 @@ CLASSROOM_CHOICES = (
     ('8D', '8D'),
 )
 
+ROLE_CHOICES = (
+    ('', ''),
+
+    ('student', 'Student'),
+    ('instructor', 'Instructor'),
+)
+
 
 class ExtraRegisterForm(ModelForm):
     """
     The fields on this form are derived from the StudentParent model
     """
-    classroom = forms.ChoiceField(label='Classroom', choices=CLASSROOM_CHOICES, initial=0)
+    role = forms.ChoiceField(label='Role', choices=ROLE_CHOICES)
+    classroom = forms.ChoiceField(label='Classroom', choices=CLASSROOM_CHOICES)
     phone = forms.CharField(label='Phone number')
     parent_phone = forms.CharField(label='Parent phone number')
     parent_email = forms.CharField(label='Parent email')
     teacher = forms.ModelChoiceField(label='Teacher', queryset=User.objects.filter(is_staff=True, is_superuser=False))
-    city = forms.CharField(label='City')
-    school = forms.CharField(label='School')
+    city = forms.ModelChoiceField(label='City', queryset=City.objects.all())
+    school = forms.ModelChoiceField(label='School', queryset=School.objects.all())
 
     class Meta(object):
         model = StudentParent
-        fields = ('phone', 'parent_email', 'parent_phone', 'teacher', 'city', 'school', 'classroom')
+        fields = ('role', 'phone', 'parent_email', 'parent_phone', 'teacher', 'city', 'school', 'classroom')
