+++ lms/templates/student_account/login_and_register.html
@@ -15,6 +15,10 @@
     <%static:require_module module_name="js/student_account/logistration_factory" class_name="LogistrationFactory">
         var options = ${data | n, dump_js_escaped_json};
         LogistrationFactory(options);
+
+        // set default student role after form loading
+        $('#register-role').val('student');
+
         if ('newrelic' in window) {
             newrelic.finished();
             // Because of a New Relic bug, the finished() event doesn't show up
@@ -27,6 +31,22 @@
     % if configuration_helpers.get_value('DISPLAY_TOS_IN_MODAL_ON_REGISTRATION_PAGE', False):
     <script type="text/javascript" src="${static.url('js/student_account/tos_modal.js')}"></script>
     % endif
+
+    <script type="text/javascript">
+        $(document).on('change', '#register-role', function() {
+            var value = this.value;
+
+            var classroomField = $('.form-field.select-classroom');
+            var parentEmailField = $('.form-field.text-parent_email');
+            var parentPhoneField = $('.form-field.text-parent_phone');
+            var teacherField = $('.form-field.select-teacher');
+
+            var studentFields = [classroomField, parentEmailField, parentPhoneField, teacherField];
+            studentFields.forEach(function(field) {
+                field.toggle(value === 'student');
+            });
+        });
+    </script>
 </%block>
 
 <%block name="header_extras">
